+++ js/main.js
@@ -1,10 +1,10 @@
-import { loadPhotosFromServer } from './data.js';
+// import { loadPhotosFromServer } from './data.js';
 import { renderThumbnails } from './thumbnails.js';
 import { openBigPicture } from './big-picture.js';
 // import { increaseScale, decreaseScale, updateScale } from './scale.js';
 // import { applyEffect, removeAllEffectClasses, updateEffectStyle } from './slider.js';
 import { hideOverlay, resetForm } from './forms.js';
-import { sendDataToServer, blockSubmitButton, unblockSubmitButton, createFormData } from './server.js';
+import { sendDataToServer, blockSubmitButton, unblockSubmitButton, createFormData,loadPhotosFromServer} from './server.js';
 
 const DATA_URL = 'https://28.javascript.htmlacademy.pro/kekstagram/data';
